Allow externally supplied keys
In some situations, keys should be generated externally, then a CSR submitted to be signed. This change allows that by looking for a CSR in the host directory, if found, the key is not generated, instead the CSR is moved into the dated cert directory and used to generate a certificate.