Add file based password store
CA password can now be stored in a file. If no password is entered, a random password will be generated into this file, which causes a warning to be printed. This is only intended to be used when no in a production setting, or if access to the entire caman installation is secured.